% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabyl_pivot_longer.R
\name{tabyl_pivot_longer}
\alias{tabyl_pivot_longer}
\title{Title}
\usage{
tabyl_pivot_longer(tabyl, col_var = NULL, list_var = NULL)
}
\arguments{
\item{tabyl}{}
}
\value{

}
\description{
Title
}
\examples{
datasets::Titanic \%>\%
  data.frame() \%>\%
  janitor::clean_names() \%>\%
  tidyr::uncount(weights = freq) \%>\%
  janitor::tabyl(survived, sex) \%>\%
  janitor::adorn_totals(c("row", "col")) \%>\%
  janitor::adorn_percentages(denominator = "col") \%>\%
  janitor::adorn_pct_formatting() \%>\%
  janitor::adorn_ns(position = "front") \%>\%
  tabyl_pivot_longer(col_var = "sex")

 datasets::Titanic \%>\%
  data.frame() \%>\%
  janitor::clean_names() \%>\%
  tidyr::uncount(weights = freq) \%>\%
  janitor::tabyl(survived, sex,class) \%>\%
  janitor::adorn_totals(c("row", "col")) \%>\%
  janitor::adorn_percentages(denominator = "col") \%>\%
  janitor::adorn_pct_formatting() \%>\%
  janitor::adorn_ns(position = "front") \%>\%
  tabyl_pivot_longer(col_var = "sex", list_var = "class")

library(ggplot2)
 datasets::Titanic \%>\%
  data.frame() \%>\%
  janitor::clean_names() \%>\%
  tidyr::uncount(weights = freq) \%>\%
  janitor::tabyl(survived, sex, class)  \%>\%
  tabyl_pivot_longer(col_var = "sex", list_var = "class") \%>\%
  ggplot() +
  aes(x = sex, y = survived) +
  facet_wrap(~class) +
  geom_tile() +
  aes(fill = value) +
  geom_text(aes(label = value), color = "white")

 datasets::Titanic \%>\%
  data.frame() \%>\%
  janitor::clean_names() \%>\%
  tidyr::uncount(weights = freq) \%>\%
  janitor::tabyl(survived, sex,class) \%>\%
  janitor::adorn_totals(c("row", "col")) \%>\%
  janitor::adorn_percentages(denominator = "col") ->
  percentages_prep

  percentages_prep \%>\%
  janitor::adorn_pct_formatting() \%>\%
  janitor::adorn_ns(position = "front") \%>\%
  tabyl_pivot_longer(col_var = "sex", list_var = "class") \%>\%
  ggplot() +
  aes(x = sex, y = survived \%>\% forcats::fct_rev()) +
  facet_wrap( ~ class) +
  geom_tile(data = percentages_prep \%>\%
  tabyl_pivot_longer(col_var = "sex", list_var = "class"),
  aes(fill = value)) +
  geom_text(aes(label = value), color = "white") +
  scale_fill_viridis_c(limits = 0:1)
}
