% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_three_way.R
\name{tabyl_three_way_flatten}
\alias{tabyl_three_way_flatten}
\title{Title}
\usage{
tabyl_three_way_flatten(tabyl, list_var = NULL)
}
\arguments{
\item{tabyl}{}
}
\value{

}
\description{
Title
}
\examples{
datasets::Titanic \%>\%
  data.frame() \%>\%
  janitor::clean_names() \%>\%
  tidyr::uncount(weights = freq) \%>\%
  janitor::tabyl(survived, class, sex) \%>\%
  janitor::adorn_totals(c("row", "col")) \%>\%
  janitor::adorn_percentages(denominator = "col") \%>\%
  janitor::adorn_pct_formatting() \%>\%
  janitor::adorn_ns(position = "front") \%>\%
  tabyl_three_way_flatten()

datasets::Titanic \%>\%
  data.frame() \%>\%
  janitor::clean_names() \%>\%
  tidyr::uncount(weights = freq) \%>\%
  janitor::tabyl(survived, class, sex) \%>\%
  janitor::adorn_totals(c("row", "col")) \%>\%
  tabyl_three_way_flatten(list_var = "sex")


}
